WIP Fix erlang dependency loading.
Work in progress rework of the dependency loading. This breaks the
contract around the library loading, and I'm still sorting that
out. Putting this out there for commentary on the general approach.

Rebar3 now provides a feature for overriding the loading of
dependencies. If a dep is provided in the _checkouts directory, rebar3
will build it in preference to the version loaded with get-deps.

This rewrite uses a soft link from _checkouts to /host/external-deps;
the dev at this time will have to have a link on the host side to
their checkout of the dep.

I'm not quite sure how the old code was supposed to work with loading
things into /host/*, as sync was pretty aggressive about stomping on
things there.

Signed-off-by: Mark Anderson <mark@chef.io>